Add Pico W/board type support
A lot of changes in this PR.

Add the ability to select one of the boards types available in the SDK
Adds a section for PicoW specific library inclusion (not yet added to the command line options)

Framework:

Changes the way parameters are passed around from a class to a map indexed on parameter names.
Removed the CHeckListBox class, and implemented the features using tthe standard checkbox instead
Added facility to enter SSID and passwords - although this is not yet used.