Add support for Raspberry Pi Debug Probe
Added the option to use a Raspberry Pi Debug Probe (cmsis-dap debugger). This required the following:

(i) an additional parameter to set the adapter speed
(ii) the interface configuration file cmsis-dap.cfg

Tested on a Pico W with Raspberry Pi Debug Probe.

For other options, serverArgs is an empty array